Move thinblock and graphene files into src/blockrelay
This is in preparation for more refactoring and re-use of common
code.
