+++ src/main.cpp
@@ -8,6 +8,8 @@
 
 #include "addrman.h"
 #include "arith_uint256.h"
+#include "blockrelay/graphene.h"
+#include "blockrelay/thinblock.h"
 #include "blockstorage/blockstorage.h"
 #include "blockstorage/sequential_files.h"
 #include "chainparams.h"
@@ -20,7 +22,6 @@
 #include "consensus/validation.h"
 #include "dosman.h"
 #include "expedited.h"
-#include "graphene.h"
 #include "hash.h"
 #include "init.h"
 #include "merkleblock.h"
@@ -36,7 +37,6 @@
 #include "script/script.h"
 #include "script/sigcache.h"
 #include "script/standard.h"
-#include "thinblock.h"
 #include "tinyformat.h"
 #include "txdb.h"
 #include "txmempool.h"
