+++ src/test/exploit_tests.cpp
@@ -2,6 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+#include "blockrelay/thinblock.h"
 #include "bloom.h"
 #include "chainparams.h"
 #include "dosman.h"
@@ -14,7 +15,6 @@
 #include "serialize.h"
 #include "streams.h"
 #include "streams.h"
-#include "thinblock.h"
 #include "txmempool.h"
 #include "uint256.h"
 #include "unlimited.h"
