+++ src/test/thinblock_tests.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "thinblock.h"
+#include "blockrelay/thinblock.h"
 #include "bloom.h"
 #include "chainparams.h"
 #include "main.h"
