+++ src/test/thinblock_util_tests.cpp
@@ -3,7 +3,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "thinblock.h"
+#include "blockrelay/thinblock.h"
 #include "test/test_bitcoin.h"
 #include <boost/test/unit_test.hpp>
 #include <limits>
