+++ src/unlimited.cpp
@@ -5,6 +5,8 @@
 #include "unlimited.h"
 
 #include "base58.h"
+#include "blockrelay/graphene.h"
+#include "blockrelay/thinblock.h"
 #include "blockstorage/blockstorage.h"
 #include "cashaddrenc.h"
 #include "chain.h"
@@ -19,7 +21,6 @@
 #include "dosman.h"
 #include "dstencode.h"
 #include "expedited.h"
-#include "graphene.h"
 #include "hash.h"
 #include "leakybucket.h"
 #include "miner.h"
@@ -31,7 +32,6 @@
 #include "rpc/server.h"
 #include "script/standard.h"
 #include "stat.h"
-#include "thinblock.h"
 #include "timedata.h"
 #include "tinyformat.h"
 #include "tweak.h"
