+++ src/unlimited.h
@@ -6,6 +6,7 @@
 #ifndef BITCOIN_UNLIMITED_H
 #define BITCOIN_UNLIMITED_H
 
+#include "blockrelay/thinblock.h"
 #include "chain.h"
 #include "checkqueue.h"
 #include "coins.h"
@@ -15,7 +16,6 @@
 #include "net.h"
 #include "script/script_error.h"
 #include "stat.h"
-#include "thinblock.h"
 #include "tweak.h"
 #include "uahf_fork.h"
 #include "univalue/include/univalue.h"
